win_requirements.txt
pycryptodome>=3.19.0
Pillow>=10.0.0
pyreadline3>=3.4.1
pywin32>=306
requests>=2.31.0
colorama>=0.4.6
python-dateutil>=2.8.2